requirements.txt
fastapi
uvicorn
sendgrid
python-dotenv
sqlalchemy
psycopg2-binary
slowapi